requirements.txt
# List required packages in this file, one per line.
rucio-clients
